Split info server into standalone module
