+++ owners/index.js
@@ -17,23 +17,18 @@
 require('dotenv').config();
 
 const Octokit = require('@octokit/rest');
+const express = require('express');
+
 const {GitHub, PullRequest, Team} = require('./src/github');
 const {LocalRepository} = require('./src/repo');
 const {OwnersBot} = require('./src/owners_bot');
 const {OwnersParser} = require('./src/parser');
 const {OwnersCheck} = require('./src/owners_check');
-const express = require('express');
 
 const GITHUB_ACCESS_TOKEN = process.env.GITHUB_ACCESS_TOKEN;
 const GITHUB_REPO = process.env.GITHUB_REPO || 'ampproject/amphtml';
 const [GITHUB_REPO_OWNER, GITHUB_REPO_NAME] = GITHUB_REPO.split('/');
-
-const APP_ID = process.env.APP_ID || 'UNKNOWN';
-const APP_COMMIT_SHA = process.env.APP_COMMIT_SHA || 'UNKNOWN';
-const APP_COMMIT_MSG = process.env.APP_COMMIT_MSG || 'UNKNOWN';
 const INFO_SERVER_PORT = Number(process.env.INFO_SERVER_PORT || 8081);
-
-const CACHED_TREE_REFRESH_MS = 10 * 60 * 1000;
 
 module.exports = app => {
   const localRepo = new LocalRepository(process.env.GITHUB_REPO_DIR);
@@ -100,71 +95,14 @@ module.exports = app => {
     }
   );
 
-  // Since the status server is publicly accessible, we don't want any
-  // endpoints to be making API calls or doing disk I/O. Rather than parsing
-  // the file tree from the local repo on every request, we keep a local copy
-  // and update it every ten minutes.
-  const parser = new OwnersParser(localRepo, ownersBot.teams, app.log);
-  let treeParse = {result: {}, errors: []};
-  /** Updates the cached copy of the parsed ownership tree. */
-  function updateTree() {
-    app.log('Updating cached owners tree');
-    parser.parseOwnersTree().then(parse => {
-      treeParse = parse;
-    });
-  }
-
   if (process.env.NODE_ENV !== 'test') {
-    updateTree();
-    teamsInitialized.then(updateTree);
-    setInterval(updateTree, CACHED_TREE_REFRESH_MS);
-
-    /** Health check server endpoints **/
-    const expressApp = express();
-    expressApp.get('/status', (req, res) => {
-      res.send(
-        [
-          `The OWNERS bot is live and running on ${GITHUB_REPO}!`,
-          `App ID: ${APP_ID}`,
-          `Deployed commit: <code>${APP_COMMIT_SHA}</code> ${APP_COMMIT_MSG}`,
-          '<a href="/tree">Owners Tree</a>',
-          '<a href="/teams">Organization Teams</a>',
-        ].join('<br>')
-      );
-    });
-
-    expressApp.get('/tree', (req, res) => {
-      const treeHeader = '<h3>OWNERS tree</h3>';
-      const treeDisplay = `<pre>${treeParse.result.toString()}</pre>`;
-
-      let output = `${treeHeader}${treeDisplay}`;
-      if (treeParse.errors.length) {
-        const errorHeader = '<h3>Parser Errors</h3>';
-        const errorDisplay = treeParse.errors
-          .map(error => error.toString())
-          .join('<br>');
-        output += `${errorHeader}<code>${errorDisplay}</code>`;
-      }
-
-      res.send(output);
-    });
-
-    expressApp.get('/teams', (req, res) => {
-      const teamSections = [];
-      Object.entries(ownersBot.teams).forEach(([name, team]) => {
-        teamSections.push(
-          [
-            `Team "${name}" (ID: ${team.id}):`,
-            ...team.members.map(username => `- ${username}`),
-          ].join('<br>')
-        );
-      });
-
-      res.send(['<h2>Teams</h2>', ...teamSections].join('<br><br>'));
-    });
-
-    expressApp.listen(INFO_SERVER_PORT, () => {
-      app.log(`Starting status server on port ${INFO_SERVER_PORT}`);
+    // Since the status server is publicly accessible, we don't want any
+    // endpoints to be making API calls or doing disk I/O. Rather than parsing
+    // the file tree from the local repo on every request, we keep a local copy
+    // and update it every ten minutes.
+    const parser = new OwnersParser(localRepo, ownersBot.teams, app.log);
+    teamsInitialized.then(() => {
+      healthServer(INFO_SERVER_PORT, parser, app.log);
     });
   }
